+++ crates/zng/src/lib.rs
@@ -837,9 +837,11 @@ mod defaults {
         /// * [`ImageManager`]
         /// * [`ClipboardManager`]
         /// * [`UndoManager`]
+        /// * [`SingleInstanceManager`] if the `"single_instance"` feature is enabled.
         /// * [`MaterialFonts`] if any `"material_icons*"` feature is enabled.
         ///
         /// [`MaterialFonts`]: zng_wgt_material_icons::MaterialFonts
+        /// [`SingleInstanceManager`]: zng_ext_single_instance::SingleInstanceManager
         /// [`ConfigManager`]: zng_ext_config::ConfigManager
         /// [`L10nManager`]: zng_ext_l10n::L10nManager
         /// [`FontManager`]: zng_ext_font::FontManager
@@ -860,6 +862,9 @@ mod defaults {
                 .extend(ImageManager::default())
                 .extend(ClipboardManager::default())
                 .extend(UndoManager::default());
+
+            #[cfg(feature = "single_instance")]
+            let r = r.extend(zng_ext_single_instance::SingleInstanceManager::default());
 
             #[cfg(any(
                 feature = "material_icons_outlined",