+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * **Breaking:** Remove `"bytemuck"` feature from `zng-unique-id`.
     - Now must use `impl_unique_id_bytemuck!` to generate the impls.
     - Note that this is only a breaking change for direct dependents of `zng-unique-id`.
+* Add single app-process instance mode.
+    - Adds `zng-ext-single-instance` crate re-exported in `zng::app` for feature `"single_instance"`.
 
 # 0.3.4